backend/requirements.txt
pika
psycopg2-binary
fastapi
uvicorn
requests

google-genai
python-dotenv
bcrypt==4.0.1

fastapi
uvicorn[standard]
psycopg2-binary
python-jose[cryptography]
passlib[bcrypt]
python-multipart
pydantic
pydantic-settings